Moving the input specific data from preprocess_input.py to formatting to wrapper.py
