+++ tsLQC/preprocess_input.py
@@ -1,21 +1,9 @@
 import pandas as pd
 import numpy as np
 from statsmodels.tsa.seasonal import seasonal_decompose
-from tsLQC.constant import primary_key
+
 
 percent_trimmed = 0.1
-
-
-def preprocessing(timeseries_input_df: pd.DataFrame) -> pd.DataFrame:
-
-    timeseries_input_df['Date'] = timeseries_input_df['Date'].str[-4:] + '-' + timeseries_input_df['Date'].str[
-                                                                               3:5] + '-01'
-    timeseries_input_df['Date'] = pd.to_datetime(timeseries_input_df['Date'])
-    timeseries_input_df['CompanyClass'] = timeseries_input_df['CompanyClass'].astype('str')
-    timeseries_input_df = timeseries_input_df.set_index(primary_key)
-    timeseries_input_df.index = timeseries_input_df.index.astype(int).astype(str)
-
-    return timeseries_input_df
 
 
 def outlier_treatment(ts: pd.Series) -> pd.Series:
